+++ R/tm_g_km.R
@@ -168,10 +168,14 @@ template_g_km <- function(dataname = "ANL",
                       sep = "\n"
                     )
                   ),
-                  footnotes = paste(
-                    "Ties for Coxph (Hazard Ratio):", ties, "\n",
-                    "p-value Method for Coxph (Hazard Ratio):", pval_method
-                  ),
+                  footnotes = if (annot_coxph) {
+                    paste(
+                      "Ties for Coxph (Hazard Ratio):", ties, "\n",
+                      "p-value Method for Coxph (Hazard Ratio):", pval_method
+                    )
+                  } else {
+                    NULL
+                  },
                   ggtheme = ggplot2::theme_minimal(),
                   annot_surv_med = annot_surv_med,
                   annot_coxph = annot_coxph,
@@ -240,10 +244,14 @@ template_g_km <- function(dataname = "ANL",
               title,
               paste(title, paste("Stratified by", paste(strata_var, collapse = ", ")), sep = "\n")
             ),
-            footnotes = paste(
-              "Ties for Coxph (Hazard Ratio):", ties, "\n",
-              "p-value Method for Coxph (Hazard Ratio):", pval_method
-            ),
+            footnotes = if (annot_coxph) {
+              paste(
+                "Ties for Coxph (Hazard Ratio):", ties, "\n",
+                "p-value Method for Coxph (Hazard Ratio):", pval_method
+              )
+            } else {
+              NULL
+            },
           )
           result
         },