+++ NEWS.md
@@ -16,6 +16,7 @@
 * Added Stratified Analysis CI method option panel to `tm_t_binary_outcome`.
 * Added validation for covariate/visit conflicts to `tm_a_mmrm`.
 * Remove unnecessary brackets from header in `tm_t_exposure`.
+* Hid footnotes in `tm_g_km` and `tm_t_tte` when "Compare Treatments" is off.
 
 ### Bug fixes
 
@@ -26,6 +27,7 @@
 * Fixed bug in `tm_t_binary_outcome` that crashed the app when deselecting all `paramcd`.
 * Fixed `teal.reporter` card names for `tm_t_smq`.
 * Fixed bug in `tm_t_shift_by_arm_by_worst` by adding validations for choosing different endpoint values.
+* Fixed bug in `tm_t_coxreg` preventing footnotes from displaying for univariate models.
 
 ### Miscellaneous
 * Added `nestcolor` dependency and replaced deprecated function `tern::color_palette` with `nestcolor::color_palette`.